Refactoring : JoinedSubclassDiscriminatorMappingImpl and EntityDiscriminatorMappingImpl
